Split bnf parsers out of data into ll1 and descent modules
Move the parser/automaton backends out of `fs/bnf/data`, leaving that
module as the pure serializable IR (`RuleSet` + `toData`). The LL(1)
dispatch/matcher now lives in `fs/bnf/ll1` and the recursive descent
matcher in `fs/bnf/descent`, each a sibling built over the data IR, so
new backends (recognizer, dfa) can land the same way.

Tests are split per module and `fs/djs/tokenizer-new` now imports
`descentParser` from `fs/bnf/descent`.
