+++ fs/bnf/data/module.f.ts
@@ -1,23 +1,24 @@
 /**
- * Provides grammar-to-data conversion and LL(1) dispatch/matching types and helpers
- * for FunctionalScript BNF rules.
+ * The serializable BNF intermediate representation (IR) and the
+ * {@link toData} conversion from the functional grammar.
+ *
+ * This module is the pure, parser-agnostic substrate: it defines the
+ * {@link RuleSet} data form (`Rule = Variant | Sequence | TerminalRange`) and
+ * converts a functional grammar into it. The automaton builders that consume a
+ * {@link RuleSet} live in their own sibling modules (`fs/bnf/ll1`,
+ * `fs/bnf/descent`, …), so the IR stays free of any one parser's machinery.
  *
  * @module
  */
-import { type CodePoint, stringToCodePointList } from '../../text/utf16/module.f.ts'
-import { strictEqual } from '../../types/function/operator/module.f.ts'
+import { stringToCodePointList } from '../../text/utf16/module.f.ts'
 import { map, toArray } from '../../types/list/module.f.ts'
-import { rangeMap, type RangeMapArray } from '../../types/range_map/module.f.ts'
-import { contains, set, type StringSet } from '../../types/string_set/module.f.ts'
-import { contains as rangeContains } from '../../types/range/module.f.ts'
 import {
     oneEncode,
-    rangeDecode,
     type DataRule,
     type Rule as FRule,
     type Sequence as FSequence,
 } from '../module.f.ts'
-import { definedEntries, type StringMap } from '../../types/object/module.f.ts'
+import { type StringMap } from '../../types/object/module.f.ts'
 
 /**
  * Encoded terminal range value used by BNF data rules.
@@ -50,100 +51,6 @@ export type RuleSet = Readonly<Record<string, Rule>>
 //
 
 type FRuleMap = StringMap<string, FRule>
-
-type EmptyTag = string|true|undefined
-
-type EmptyTagEntry = string|boolean
-
-type DispatchRule = {
-    readonly emptyTag: EmptyTag,
-    readonly rangeMap: Dispatch
-}
-
-type Dispatch = RangeMapArray<DispatchResult>
-
-type DispatchResult = DispatchRuleCollection | null
-
-type DispatchRuleOrName = DispatchRule | string
-
-type DispatchRuleCollection = {
-    readonly tag: string | undefined,
-    readonly rules: DispatchRuleOrName[]
-}
-
-type DispatchMap = StringMap<string, DispatchRule>
-
-type EmptyTagMap = StringMap<string, EmptyTagEntry>
-
-/**
- * Recursive descent matcher for a single named rule.
- */
-export type DescentMatchRule<T> = (name: string, tag: AstTag, s: readonly CodePointMeta<T>[], idx: number) => DescentMatchResult<T>
-
-/**
- * Result tuple of a descent match operation: AST node, success flag, and next index.
- */
-export type DescentMatchResult<T> = readonly[AstRuleMeta<T>, boolean, number]
-
-/**
- * Entry-point recursive descent matcher.
- */
-export type DescentMatch<T> = (name: string, s: readonly CodePointMeta<T>[]) => DescentMatchResult<T>
-
-/**
- * Code point value paired with metadata.
- */
-export type CodePointMeta<T> = readonly[CodePoint, T]
-
-/**
- * AST sequence for the metadata-aware parser.
- */
-export type AstSequenceMeta<T> = readonly(AstRuleMeta<T>|CodePointMeta<T>)[]
-
-/**
- * Metadata-aware AST node.
- */
-export type AstRuleMeta<T> = {
-    readonly tag: AstTag,
-    readonly sequence: AstSequenceMeta<T>
-}
-
-/**
- * Represents a parsed Abstract Syntax Tree (AST) sequence.
- */
-export type AstSequence = readonly(AstRule|CodePoint)[]
-
-export type AstTag = string|true|undefined
-
-/**
- * Represents a parsed AST rule, consisting of a rule name and its parsed sequence.
- */
-type AstRule = {
-    readonly tag: AstTag,
-    readonly sequence: AstSequence
-}
-
-/**
- * Represents the remaining input after a match attempt, or `null` if no match is possible.
- */
-export type Remainder = readonly CodePoint[] | null
-
-/**
- * Parsing result of {@link parser} and {@link parserRuleSet}.
- *
- * Represents the result of a match operation, including the parsed AST rule and the remainder of the input.
- */
-export type MatchResult = readonly[AstRule, boolean, Remainder]
-
-/**
- * LL(1) parser function for matching by rule name.
- */
-export type Match = (name: string, s: readonly CodePoint[]) => MatchResult
-
-/**
- * Internal match function signature used by compiled dispatch rules.
- */
-export type MatchRule = (dr: DispatchRule, s: readonly CodePoint[]) => MatchResult
 
 const { entries } = Object
 
@@ -232,256 +139,4 @@ const toDataAdd = (map: FRuleMap) => (fr: FRule): readonly [FRuleMap, RuleSet, s
 export const toData = (fr: FRule): readonly [RuleSet, string] => {
     const [, ruleSet, id] = toDataAdd({})(fr)
     return [ruleSet, id]
-}
-
-const dispatchOp = rangeMap<DispatchResult>({
-    union: a => b => {
-        if (a === null) {
-            return b
-        }
-        if (b === null) {
-            return a
-        }
-        throw ['can not merge [', a, '][', b, ']']
-    },
-    equal: strictEqual,
-    def: null,
-})
-
-/**
- * Builds a dispatch map for a {@link RuleSet} to enable predictive parsing.
- */
-export const dispatchMap = (ruleSet: RuleSet): DispatchMap => {
-
-    const addRuleToDispatch = (dr: DispatchResult, name: string): DispatchResult => {
-        if (dr === null)
-            return null
-
-        return { tag: dr.tag, rules: [...dr.rules, name]}
-    }
-
-    const addTagToDispatch = (dr: DispatchResult, tag: string): DispatchResult => {
-        if (dr === null)
-            return null
-
-        return { tag, rules: dr.rules}
-    }
-
-    const dispatchRule = (dm: DispatchMap, name: string, current: StringSet): DispatchMap => {
-        if (name in dm) { return dm }
-        const newCurrent = set(name)(current)
-        const rule = ruleSet[name]
-        if (typeof rule === 'number') {
-            const range = rangeDecode(rule)
-            const dispatch = dispatchOp.fromRange(range)({tag: undefined, rules: []})
-            const dr: DispatchRule = {emptyTag: undefined, rangeMap: dispatch}
-            return { ...dm, [name]: dr }
-        } else if (rule instanceof Array) {
-            let emptyTag: EmptyTag = true
-            let result: Dispatch = []
-            for (const item of rule) {
-                if (contains(item)(newCurrent)) {
-                    result = result.map(x => [addRuleToDispatch(x[0], item), x[1]])
-                } else {
-                    dm = dispatchRule(dm, item, newCurrent)
-                    const dr = dm[item]!
-                    if (emptyTag === true) {
-                        result = result.map(x => [addRuleToDispatch(x[0], item), x[1]])
-                        result = toArray(dispatchOp.merge(result)(dr.rangeMap))
-                        emptyTag = dr.emptyTag !== undefined ? true : undefined
-                    } else {
-                        result = result.map(x => [addRuleToDispatch(x[0], item), x[1]])
-                    }
-                }
-            }
-            const dr: DispatchRule = {emptyTag, rangeMap: result}
-            return { ...dm, [name]: dr}
-        } else {
-            const entries = definedEntries(rule)
-            let result: Dispatch = []
-            let emptyTag: EmptyTag = undefined
-            for (const [tag, item] of entries) {
-                dm = dispatchRule(dm, item, newCurrent)
-                const dr = dm[item]!
-                if (dr.emptyTag !== undefined) {
-                    emptyTag = tag
-                } else {
-                    const d: Dispatch = dr.rangeMap.map(x => [addTagToDispatch(x[0], tag), x[1]])
-                    result = toArray(dispatchOp.merge(result)(d))
-                }
-            }
-            const dr: DispatchRule = {emptyTag, rangeMap: result}
-            return { ...dm, [name]: dr}
-        }
-    }
-
-    let result: DispatchMap = {}
-    for (const k in ruleSet) {
-        result = dispatchRule(result, k, null)
-    }
-
-    return result
-}
-
-const emptyTagMapAdd = (ruleSet: RuleSet) => (map: EmptyTagMap) => (name: string): readonly [RuleSet, EmptyTagMap, EmptyTagEntry] => {
-    if (name in map) {
-        return [ruleSet, map, map[name]!]
-    }
-
-    const rule = ruleSet[name]
-
-    if (typeof rule === 'number') {
-        return [ruleSet, { ...map, [name]: false }, false]
-    } else if (rule instanceof Array) {
-        map = { ...map, [name]: true}
-        let emptyTag: EmptyTagEntry = rule.length == 0
-        for (const item of rule) {
-            const [,newMap,itemEmptyTag] = emptyTagMapAdd(ruleSet)(map)(item)
-            map = newMap
-            if (emptyTag === false) {
-                emptyTag = itemEmptyTag !== false
-            }
-        }
-        return [ruleSet, { ...map, [name]: emptyTag }, emptyTag]
-    } else {
-        map = { ...map, [name]: true}
-        const entries = definedEntries(rule)
-        let emptyTag: EmptyTagEntry = false
-        for (const [tag, item] of entries) {
-            const [,newMap,itemEmptyTag] = emptyTagMapAdd(ruleSet)(map)(item)
-            map = newMap
-            if (itemEmptyTag !== false) {
-                emptyTag = tag
-            }
-        }
-        return [ruleSet, { ...map, [name]: emptyTag }, emptyTag]
-    }
-}
-
-/**
- * Creates a map that describes whether each rule can consume empty input and,
- * for tagged variants, which tag represents the empty match.
- */
-export const createEmptyTagMap = (data: readonly [RuleSet, string]): EmptyTagMap => {
-    return emptyTagMapAdd(data[0])({})(data[1])[1]
-}
-
-/**
- * Creates a recursive descent parser that preserves metadata for each consumed
- * code point.
- */
-export const descentParser = <T>(fr: FRule): DescentMatch<T> => {
-    const data = toData(fr)
-    const emptyTagMap = createEmptyTagMap(data)
-
-    const getEmptyTag = (name: string): EmptyTag => {
-        const res = emptyTagMap[name]
-        return res === false ? undefined : res
-    }
-
-    const f: DescentMatchRule<T> = (name, tag, cp, idx): DescentMatchResult<T> => {
-        const mrSuccess = (tag: AstTag, sequence: AstSequenceMeta<T>, idx: number): DescentMatchResult<T> => [{tag, sequence}, true, idx]
-        const mrFail = (tag: AstTag, sequence: AstSequenceMeta<T>, idx: number): DescentMatchResult<T> => [{tag, sequence}, false, idx]
-
-        const rule = data[0][name]
-        if (typeof rule === 'number') {
-            const emptyTag = getEmptyTag(name)
-            if (idx >= cp.length) {
-                return emptyTag === undefined ? mrFail(emptyTag, [], idx) : mrSuccess(emptyTag, [], idx)
-            }
-
-            const cpi = cp[idx]
-            const range = rangeDecode(rule)
-            if (rangeContains(range)(cpi[0])) {
-                return mrSuccess(tag, [cpi], idx + 1)
-            }
-            return mrFail(emptyTag, [], idx)
-        } else if (rule instanceof Array) {
-            let seq: AstSequenceMeta<T> = []
-            let tidx = idx
-            for (const item of rule) {
-                const m = f(item, undefined, cp, tidx)
-                const [astRule, success, nidx] = m
-                tidx = nidx
-                if (success === false) {
-                    return mrFail(tag, [], idx)
-                }
-                seq = [...seq, astRule]
-            }
-            return mrSuccess(tag, seq, tidx)
-        } else {
-            const entries = definedEntries(rule)
-            const emptyTag = getEmptyTag(name)
-            let emptyResult = mrFail(emptyTag, [], idx)
-            for (const [tag, item] of entries) {
-                const m = f(item, tag, cp, idx)
-                if (m[1]) {
-                    if (idx !== m[2])
-                        return m
-
-                    emptyResult = m
-                }
-            }
-            return emptyResult
-        }
-    }
-
-    const match: DescentMatch<T> = (name, cp): DescentMatchResult<T> => {
-        return f(name, undefined, cp, 0)
-    }
-
-    return match
-}
-
-/**
- * Creates an LL(1) parser from a functional grammar rule.
- */
-export const parser = (fr: FRule): Match => {
-    const data = toData(fr)
-    return parserRuleSet(data[0])
-}
-
-/**
- * Creates an LL(1) parser from an already materialized {@link RuleSet}.
- */
-export const parserRuleSet = (ruleSet: RuleSet): Match => {
-    const map = dispatchMap(ruleSet)
-
-    const f: MatchRule = (rule, cp): MatchResult => {
-        const mrSuccess = (tag: AstTag, sequence: AstSequence, r: Remainder): MatchResult => [{tag, sequence}, true, r]
-        const mrFail = (tag: AstTag, sequence: AstSequence, r: Remainder): MatchResult => [{tag, sequence}, false, r]
-        const {emptyTag, rangeMap} = rule
-        if (cp.length === 0) {
-            return mrSuccess(emptyTag, [], emptyTag === undefined ? null : cp)
-        }
-        const cp0 = cp[0]
-        const dr = dispatchOp.get(cp0)(rangeMap)
-        if (dr === null) {
-            if (emptyTag === undefined) {
-                return mrFail(emptyTag, [], cp)
-            }
-            return mrSuccess(emptyTag, [], cp)
-        }
-        let seq: AstSequence = [cp0]
-        let r = cp
-        const [_, ...restCp] = cp
-        r = restCp
-        const {tag, rules} = dr
-        for (const i of rules) {
-            const rule = typeof i === 'string' ? map[i]! : i
-            const res = f(rule, r)
-            const [astRule, success, newR] = res
-            if (success === false) {
-                return res
-            }
-            seq = [...seq, astRule]
-            if (newR === null) {
-                return mrSuccess(tag, seq, null)
-            }
-            r = newR
-        }
-        return mrSuccess(tag, seq, r)
-    }
-
-    return (name, cp): MatchResult => f(map[name]!, cp)
 }
